feat: DocumentService module - Added DocumentService module. - Moved VerifyPersonalDetails as component of DocumentService.
