+++ packages/wallets/src/features/accounts/modules/DocumentService/components/VerifyPersonalDetails/index.ts
@@ -1,0 +1,1 @@
+export { default as VerifyPersonalDetails } from './VerifyPersonalDetails';
