+++ packages/wallets/src/features/accounts/modules/DocumentService/components/index.ts
@@ -1,2 +1,1 @@
-export * from './Poi';
 export * from './VerifyPersonalDetails';
