+++ packages/wallets/src/features/accounts/modules/DocumentService/index.ts
@@ -1,0 +1,1 @@
+export { default as DocumentService } from './DocumentService';
