--- packages/wallets/src/features/accounts/modules/VerifyPersonalDetails/index.ts
@@ -1,1 +1,0 @@
-export { default as VerifyDocumentDetails } from './VerifyDocumentDetails';
